backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sendgrid==6.11.0
python-dotenv==1.0.1
Pillow==10.3.0
aiofiles==23.2.1
slowapi==0.1.9
